Moved Simformer pieces from simformer-dev to sbi codebase
